pyproject.toml
[tool.poetry]
name = "laba-med"
version = "0.1.0"
description = ""
authors = ["NovikovaAnastasia <asnovikov87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.14"
opencv-python = "^4.10.0.84"
mediapipe = "^0.10.18"
matplotlib = "^3.9.2"
ttkthemes = "^3.2.2"
pyside6 = "^6.8.1"
python-vlc = "^3.0.21203"


[tool.poetry.group.dev.dependencies]
ruff = "^0.8.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
